+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# Version 5.6.0
+* Added better support of the no power points setting
+* Bolt mods no longer show up on minor bolt
+* PP management no longer shows epic mods if the actor does not have the edge
+
 # Version 5.5.0
 * Updated the gang up calculation to match the system
 * Added pre-release support for the new gang up flags